Improving vignette/plots
Apologies for the state of this PR - I started worked on the vignette, discovered that some plots didn't work, and then in did a bunch of refactoring.

By and large the plots should look the same as before, although I've considerable simplified the underlying code and made a few minor tweaks here and there. The major difference is that sessions in warm-up/cool-down are no longer coloured differently. There's no straightforward way to do this, and I think it only worked previously due to a bug in manual scales.

I still need to finish the vignette, but I wanted to capture my progress so far.